### 2018-12-13 jschnab/pyms#2 @johnlees
Tag a release
This is required for JOSS (and the version should match that in the review 0.8.1) (see https://github.com/openjournals/joss-reviews/issues/1125)
