### 2018-12-13 jschnab/pyms#3 @johnlees
Check references in paper
openjournals/joss-reviews#1125

**References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?**

These look ok to me, but could you confirm whether DOIs are available for the references, and add them if they are?

You may wish to add one reference to either an example of where pyms has been used (e.g. to generate plots for a paper) or one to an example dataset of quantitative analysis of mycorrhizal fungi colonization of plant roots.
I would suggest adding both the references from the `README.md` to the paper too
