### 2018-12-13 jschnab/pyms-plot#4 @johnlees
Add to pip/conda
https://github.com/openjournals/joss-reviews/issues/1125
**Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.**

I would recommend adding this package to either PyPI or anaconda (as you prefer). You might like to follow these instructions for PyPI/pip:
https://gehrcke.de/2014/02/distributing-a-python-command-line-application/
or try bioconda:
https://bioconda.github.io/contributing.html

The package meets the stated requirement in JOSS, but if you are able to add to one of these sources and automatically handle dependencies at the same time I think that would be preferable
