### 2018-12-13 jschnab/pyms#5 @johnlees
Add expected output to docs
https://github.com/openjournals/joss-reviews/issues/1125
**Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?**

For the docs (either in the `README.md` or on readthedocs) I think you should add at least one image showing the expected output for the test case example
