### 2018-12-13 jschnab/pyms#6 @johnlees
Add further method documentation
https://github.com/openjournals/joss-reviews/issues/1125

**Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?**

There is a description of the statistical tests in the paper, but I believe it would be helpful to add these descriptions to either the `README.md` or readthedocs
