### 2018-12-13 #1125 @johnlees
@jschnab I've given this a go and confirmed it works with the test data. I've opened a few specific issues that will need to be addressed to meet the requirements for JOSS. My suggestion of making this into a package would definitely help with the stated goal of making this analysis easier for people who don't use the command line (typing `pip install pyms` is ok, but needing to explicitly do the same for the dependencies and cloning the github may put people off).

Some comments on the software/functionality you may also wish to take into account:

- It would be easier to specify the output directory for the statistical tests and saving the plot (also, this will be required if the input directory is not writable).
- Rather than needing to click 'analyze csv file' and find the input file again every time the settings (colour etc) are changed, it would probably be better to automatically redraw the plot when one of the menu options is chosen.
- When choosing the samples to perform statistics on, a 'select all' option might be helpful.

I also get the following message on exit:
```
Traceback (most recent call last):
  File "pyms.py", line 592, in <module>
    app.master.destroy()
  File "/Users/leesj01/anaconda3/envs/limix/lib/python3.6/tkinter/__init__.py", line 2062, in destroy
    self.tk.call('destroy', self._w)
_tkinter.TclError: can't invoke "destroy" command: application has been destroyed
```

Otherwise, looks good!
