### 2018-12-13 jschnab/pyms-plot#5 @johnlees
You should also add an example of the statistical test output, and explain what the columns are
