### 2018-12-13 jschnab/pyms-plot#6 @johnlees
Also please explain briefly in the docs which statistical tests should be used in which circumstances
