### 2018-12-14 #1125 @aokad
@jschnab I think that the following content is missing.

 1. Supporting Python versions (pyms will not work in python 2.7)
 2. "tkinter" in the dependency packages.
 3. Required version of pandas.
 I get the following error on "Analyze csv file" choose "test_colonization.csv". The isna alias for isnull was only added in version 0.21 (0.23 is the latest release)

```
Exception in Tkinter callback
Traceback (most recent call last):
  File "C:\WinPython-64bit-3.5.3.0Qt5\python-3.5.3.amd64\lib\tkinter\__init__.py", line 1559, in __call__
    return self.func(*args)
  File "pyms.py", line 351, in process_csv
    Global.col = Global.col.loc[~pd.isna(Global.col[index_name])]
AttributeError: module 'pandas' has no attribute 'isna'
```
