### 2018-12-14 #1125 @aokad
@jschnab I get the following error on "Save as" choose the file type.

```
Exception in Tkinter callback
Traceback (most recent call last):
  File "C:\WinPython-64bit-3.5.3.0Qt5\python-3.5.3.amd64\lib\tkinter\__init__.py", line 1559, in __call__
    return self.func(*args)
  File "pyms.py", line 210, in saveas_pdf
    plt.savefig(fname=fi_name, bbox_inches='tight')
  File "C:\WinPython-64bit-3.5.3.0Qt5\python-3.5.3.amd64\lib\site-packages\matplotlib\pyplot.py", line 697, in savefig
    res = fig.savefig(*args, **kwargs)
  File "C:\WinPython-64bit-3.5.3.0Qt5\python-3.5.3.amd64\lib\site-packages\matplotlib\figure.py", line 1572, in savefig
    self.canvas.print_figure(*args, **kwargs)
TypeError: print_figure() missing 1 required positional argument: 'filename'
```
