### 2019-01-12 0f8ce4c
update pyms.py: remove error when close window, choose where files are saved
