### 2019-01-12 06c0711
add 'select all' button to select all samples during statistical tests
