### 2019-01-13 52bc504
add articles from README.md to paper.bib and paper.md
