+++ paper/paper.bib
@@ -10,8 +10,9 @@
 @InProceedings{ SciPy,
   author    = { Eric Jones and Travis Oliphant and Pearu Peterson and others },
   title     = { {Scipy}: Open source scientific tools for {Python} },
-  year      = { 2001 },
-  url       = "http://www.scipy.org/"
+  year      = { 2001-- },
+  url       = "http://www.scipy.org/",
+  note      = {[Online; accessed 2019-01-13]}
 }
 
 @article{ matplotlib,
@@ -23,4 +24,24 @@
   Volume    = { 9 },
   Year      = { 2007 },
   doi       = { 10.1109/MCSE.2007.55 }
+}
+
+@article{ paszkowski,
+	Author = { Paszkowski, U., Jakovleva, L., Boller, T. },
+	Journal = { Plant Journal },
+	Pages = { 165 - 173 },
+	Title = { Maize mutants affected at distinct stages of the arbuscular mycorrhizal symbiosis },
+	Volume = { 47 },
+	Year = { 2006 },
+	doi = { 10.1111/j.1365-313X.2006.02785.x }
+}
+
+@article{ chiu,
+	Author = { Chiu, C., Choi, J., Paszkowski, U. },
+	Journal = { New Phytologist },
+	Pages = { 552 - 557 },
+	Title = { Independant signalling cues underpin arbuscular mycorrhizal symbiosis and large lateral root induction in rice },
+	Volume = { 217 },
+	Year = { 2018},
+	doi = { 10.1111/nph.14936 }
 }
