### 2019-01-13 jschnab/pyms-plot#2 @jschnab
@johnlees  - Hi, sorry but I don't understand what I should do, exactly. Can you explain how I can tag a release?

Edit: I think I figured out, please let me know if it's OK. After having a look at other projects published in JOSS, I will add another release once I have a PyMS version accepted for publication.
