### 2019-01-14 cac0cb0
update README.md and docs/index.rst
