+++ README.md
@@ -29,6 +29,10 @@ The input file should be a csv file with a specific formatting
 
 * Click on the "File" menu and point on "Save as" to display the file type choice. Click on the file extension name to save your file in the same folder as your csv file. If you generate several figures by clicking on "Analyze csv file" several times, it will save the last figure generated.
 
+Here is the expected output if you use the file *test_colonization.csv*.
+
+<img src="test_colonization.png" width="800" alt="Banner">
+
 ## Statistical analysis
 PyMS allows you to perform statistical testing of inter-genotype comparisons using the Mann-Whitney test or the Kruskal-Wallis test followed by the Dunn test for post-hoc analysis (i.e. the Kruskal-Wallis test tells you if at least one genotype is different from the others while the Dunn test tells you which genotypes are different from the others). If you want to compare two genotypes, use the Mann-Whitney test. If you want to compare multiple genotypes, use the Kruskal-Wallis then Dunn tests.
 
@@ -37,6 +41,14 @@ pop-up and you can browse your file.
 
 * Click on the "Statistics" menu and select the test you want by ticking one of 
 them. You then need to select the genotypes you want to analyze in the "Select samples" menu. You can detach the menu by clicking on the top -------, so that you don't have to open the menu each time you select a genotype. Once you selected the genotypes of interest, click on "Perform statistical test" in the "Process" menu. The p-values returned by the test will be saved in a text file in the same folder as your csv file.
+
+Here is an example if you perform the Mann-Whitney test on the genotypes *117KO* (loss-of-function mutant) and *117REV* (wild type) from the file *test_colonization.csv*. The column *Genotypes* indicates the genotypes which are compared. The other columns represent the different fungal structures. The numbers in each column represent the p-value for the difference in colonization levels for the different fungal structures.
+
+<img src="mann_whitney.png" width="800" alt="Banner">
+
+Here is another example if you perform the Kruskal-Wallis followed by Dunn tests on the genotypes *hebiba (AOC)* (loss-of-function mutant), NH WT (wild type) and PGC82-2 (complemented loss-of-function mutant). The results are displayed as for the Mann-Whitney test except there are more genotypes comparisons.
+
+<img src="kw_dunn.png" width="800" alt="Banner">
 
 ## Support
 
