### 2019-01-14 16cf6e1
correct typo in README.md
