### 2019-01-14 12a6c5b
update readthedocs with images
