### 2019-01-14 7500239
update docs with link to images on GitHub
