### 2019-01-14 jschnab/pyms-plot#2 @johnlees
Looks good to me now!
