+++ README.md
@@ -31,7 +31,7 @@ The input file should be a csv file with a specific formatting
 
 Here is the expected output if you use the file *test_colonization.csv*.
 
-<img src="test_colonization.png" width="800" alt="Banner">
+<img src="docs/images/test_colonization.png" width="800" alt="plot">
 
 ## Statistical analysis
 PyMS allows you to perform statistical testing of inter-genotype comparisons using the Mann-Whitney test or the Kruskal-Wallis test followed by the Dunn test for post-hoc analysis (i.e. the Kruskal-Wallis test tells you if at least one genotype is different from the others while the Dunn test tells you which genotypes are different from the others). If you want to compare two genotypes, use the Mann-Whitney test. If you want to compare multiple genotypes, use the Kruskal-Wallis then Dunn tests.
@@ -44,11 +44,11 @@ them. You then need to select the genotypes you want to analyze in the "Select s
 
 Here is an example if you perform the Mann-Whitney test on the genotypes *117KO* (loss-of-function mutant) and *117REV* (wild type) from the file *test_colonization.csv*. The column *Genotypes* indicates the genotypes which are compared. The other columns represent the different fungal structures. The numbers in each column represent the p-value for the difference in colonization levels for the different fungal structures.
 
-<img src="mann_whitney.png" width="800" alt="Banner">
+<img src="docs/images/mann_whitney.png" width="800" alt="mann-whitney">
 
 Here is another example if you perform the Kruskal-Wallis test followed by the Dunn tests on the genotypes *hebiba (AOC)* (loss-of-function mutant), NH WT (wild type) and PCG82-2 (complemented loss-of-function mutant). The results are displayed as for the Mann-Whitney test except there are more genotypes comparisons.
 
-<img src="kw_dunn.png" width="800" alt="Banner">
+<img src="docs/images/kw_dunn.png" width="800" alt="kruskal-wallis">
 
 ## Support
 
