### 2019-02-10 39c0075
update paper with scientific articles
