+++ paper/paper.md
@@ -19,6 +19,8 @@ bibliography: paper.bib
 
 Python for Mycorrhizal Symbiosis (PyMS) aims at providing an easier way for researchers and students to analyze quantitification of mycorrhizal symbiosis root colonization levels and produce publication-quality figures without having to use command line-based softwares. PyMS is routinely used in our laboratory for various research projects.
 
+PyMS was developed to process data collected via the line intersect method [@Paszkowski] or an equivalent method. PyMS accepts csv files and is provided with a file example from [@Chiu].
+
 PyMS allows users to plot data according to the representation most widely used (bar-plot showing the median of data and scatter-plot displaying individual data points) and select a few graphical parameters (two different modes of data grouping and bar face color including a color-blind-friendly palette). Figures can be saved in different image formats depending on downstream application.
 
 To test the hypothesis of different mycorrhizal colonization levels between groups, PyMS uses the Mann-Whitney test or its own implementation of the Kruskal-Wallis test with post-hoc analysis using the Dunn's test. These tests were chosen because the small sample sizes obtained during mycorrhizal colonization of roots makes less reliable the use of parametric tests.
