### 2019-02-11 jschnab/pyms-plot#5 @jschnab
I added images showing the expected outputs (plot and statistical tests) in both README.md and readthedocs.
