### 2019-02-11 #1125 @jschnab
@aokad Hello, I update README.md to specify that PyMS is written in Python 3, the required version of pandas (0.23) and that TkInter is required (I thought it was "vanilla" Python). The error happening when print_figure() is called is solved when pandas 0.23 is used (cf your 2nd comment).
