### 2019-02-11 jschnab/pyms-plot#6 @jschnab
I added a description and quick guide of the statistical tests to the README.md and readthedocs.
