### 2019-02-11 jschnab/pyms-plot#3 @jschnab
For the python libraries references, I followed [these guidelines](https://www.scipy.org/citing.html).

So far, PyMS was not used in a publication that was submitted to a journal, I will update the GitHub repository when this happens.

I added the two articles referenced in the `README.md` to the paper, the one from 2018 contains an example of a quantitative analysis which can be performed using PyMS.
