### 2019-02-11 #1125 @jschnab
@johnlees Hello, I added the possibility to select the output directory, I added a "select all" function for genotypes in the statistical analysis, I corrected the error when closing the window.

Now I'm working on the automatic plot re-drawing after modification of graphical parameters, and then I'll make a conda package.

Thank you all for your patience.
