+++ pyms/pyms.py
@@ -196,10 +196,15 @@ class MenuBar(Frame):
                   (.7, .13, .13)))
 
     def changeCol(self):
-       self.col = []
-       c = self.color.get()
-       for i in self.palettes[c]:
-           self.col.append(i)
+        self.col = []
+        c = self.color.get()
+        for i in self.palettes[c]:
+            self.col.append(i)
+
+         # redraw figure if it exists
+        if plt.fignum_exists(Global.figure.number):
+            plt.close('all')
+            boss.plot_data(self)
 
 class Application(Frame):
     """main application"""
@@ -391,7 +396,7 @@ indicated as column headers.\n\n')
 
         return file_name, col, geno
 
-    def process_data(self):
+    def process_data(self, geno):
         """Return number of repetitions per genotype, names of fungal 
         structures and calculate medians."""
         
@@ -414,6 +419,186 @@ indicated as column headers.\n\n')
             return number_rep, struct, medians
 
         # group by structure
+        if self.mBar.group_by.get() == 1: 
+            #count the number of repetitions for each genotype
+            number_rep = Global.col.index.value_counts()
+            number_rep = number_rep.reindex(geno)
+
+            #collect names of fungal structures
+            Global.struct = Global.col.columns[:]
+
+            #calculate medians and invert axes of dataframe
+            col_group = Global.col.groupby(Global.col.index) # group data by genotype
+            medians = col_group.median()
+            medians_trans = medians.transpose()
+
+            return number_rep, struct, medians_trans
+
+    def plot_data(self):
+        """Display bar + scatter plot."""
+
+        # group by genotype
+        if self.mBar.group_by.get() == 0:
+            #set positions and width for bars
+            pos = list(range(len(Global.geno)))
+            width = 1 / (len(Global.struct) + 1)
+
+            #plotting bars
+            Global.figure = plt.figure(figsize=(10, 5))
+            Global.ax = Global.figure.add_axes([0.1, 0.1, 0.75, 0.85])
+
+            #draw bar plot from medians
+            colors = self.mBar.col
+
+            s = 0
+            while s < len(Global.struct):
+                plt.bar([p + s * width for p in pos], Global.medians[Global.struct[s]], width,
+                        alpha=1, color=colors[s], edgecolor='black',
+                        linewidth=.5, zorder=3)
+                s += 1
+                #add legend
+                Global.ax.legend(Global.struct, bbox_to_anchor=(1, 0.5),
+                                 loc='center left')
+
+            #draw scatter plot from individual data points
+            #first you need a list with data coordinates from the 'col' DataFrame
+            start = [0]
+            end = [number_rep[0] - 1]
+            counter = 1
+            while counter < len(Global.geno):
+                start.append(end[counter - 1] + 1)
+                end.append(end[counter -1] + Global.number_rep[counter])
+                counter += 1
+
+            #browse data and add individual points on the graph
+            s = 0
+            while s < len(Global.struct):
+                i = 0
+                while i < len(Global.geno):
+                    for k in range(start[i], end[i] + 1):
+                        plt.plot(i + s * width, Global.col[Global.struct[s]].values[k],
+                                 c='black', marker='o', fillstyle='full',
+                                 zorder=4, mew=.5, markersize=1.5,
+                                 clip_on=False)
+                    i += 1
+                s += 1
+
+            #set y axis label
+            Global.ax.set_ylabel('Total root length colonization (%)')
+
+            #set position of x ticks but no display
[…]
