### 2019-02-12 70a437c
replot if figure is open when changing parameters
