### 2019-02-12 f071cf2
cleaned process_csv()
