+++ pyms/pyms.py
@@ -76,6 +76,7 @@ class Global(object):
     col = None
     number_rep = None
     medians = None
+    medians_trans = None
 
     # figure data
     figure = None
@@ -200,7 +201,7 @@ class MenuBar(Frame):
         if Global.figure:
             if plt.fignum_exists(Global.figure.number):
                 plt.close('all')
-                app.process_data()
+                Global.number_rep, Global.struct, Global.medians = app.process_data()
                 app.plot_data()
 
     def changeCol(self):
@@ -413,35 +414,35 @@ indicated as column headers.\n\n')
         if self.mBar.group_by.get() == 0:
             # count the number of repetitions for each genotype
             number_rep = Global.col.index.value_counts()
-            Global.number_rep = number_rep.reindex(Global.geno)
+            number_rep = number_rep.reindex(Global.geno)
 
             # collect names of fungal structures
-            Global.struct = Global.col.columns[:]
+            struct = Global.col.columns[:]
 
             # group data by genotype
             col_group = Global.col.groupby(Global.col.index) 
 
             # calculate medians
             medians = col_group.median()
-            Global.medians = medians.reindex(Global.geno)
+            medians = medians.reindex(Global.geno)
 
-            #return number_rep, struct, medians
+            return number_rep, struct, medians
 
         # group by structure
         elif self.mBar.group_by.get() == 1: 
             #count the number of repetitions for each genotype
             number_rep = Global.col.index.value_counts()
-            Global.number_rep = number_rep.reindex(Global.geno)
+            number_rep = number_rep.reindex(Global.geno)
 
             #collect names of fungal structures
-            Global.struct = Global.col.columns[:]
+            struct = Global.col.columns[:]
 
             #calculate medians and invert axes of dataframe
             col_group = Global.col.groupby(Global.col.index) # group data by genotype
             medians = col_group.median()
-            Global.medians_trans = medians.transpose()
+            medians_trans = medians.transpose()
 
-            #return number_rep, struct, medians_trans
+            return number_rep, struct, medians_trans
 
     def plot_data(self):
         """Display bar + scatter plot."""
@@ -529,7 +530,7 @@ indicated as column headers.\n\n')
 
             s = 0
             while s < len(Global.geno):
-                plt.bar([p + s * width for p in pos], Global.medians_trans[Global.geno[s]],
+                plt.bar([p + s * width for p in pos], Global.medians[Global.geno[s]],
                         width, alpha=1, color=colors[s], edgecolor='black',
                         linewidth=.5, zorder=3)
                 s += 1
@@ -602,7 +603,7 @@ indicated as column headers.\n\n')
         # button to select/deselect all samples
         self.mBar.me6.add_checkbutton(label='Select all', command=self.select_all_samples, variable=self.mBar.select_all, onvalue=1, offvalue=0)
         self.mBar.samples = {} #re-initialize dictionary
-        for (v, lab) in list(enumerate(geno)):
+        for (v, lab) in list(enumerate(Global.geno)):
             self.mBar.samples['{0}'.format(lab)] = IntVar()
             self.mBar.me6.add_checkbutton(label=lab, variable=self.mBar.samples[lab], command=None, onvalue=1, offvalue=0)
         self.mBar.me5.add_cascade(label='Select samples', menu=self.mBar.me6)
@@ -616,261 +617,13 @@ combination"""
         
         Global.export_number_fig += 1
 
-        if self.mBar.group_by.get() == 0:    #group by genotype
-            #=== PREPARATION ===#
-            #browse file name
-            Global.file_name = filedialog.askopenfilename()
+        Global.filename, Global.col, Global.geno = self.read_csv()
 
-            #read the csv file
-            Global.col = pd.read_csv(Global.file_name)
[…]
