+++ build/lib/pyms/pyms.py
@@ -1,0 +1,635 @@
+#script which generates a barplot from colonization data
+#written by Jonathan Schnabel, December 2017 (last update October 2018)
+#licence GPLv3
+
+#TO DO: change window icon
+#TO DO: put error handlers
+#TO DO: if a plot has already been generated, redraw figure when user
+#       changes 'group by' or 'color'
+#       can detect if plot is open with plt.fignum_exists(fig.number)
+
+### --- history --- ###
+# 15/02/2018: addition of number_rep.reindex() to re-establish original data order
+
+# 24/02/2018: addition of Mann-Whitney test
+
+# 25/02/2018: test of "pad" in xlim for up to 9 genotypes with group by genotype
+#             addition of colors to group by structure with up to 9 genotypes
+#             test with 3 structures for up to 9 genotypes
+#             test with 3 structures from 5 to 9 genotypes and group by structure
+
+# 01/03/2018: Mann-Whitney test in Python 3.5.2 does not calculate exact p-value
+#             but uses normal distribution approximation, hence n > 20 is required
+#             addition of calculation of exact p-value (mannwhitney_exact in scipy.stats)
+
+# 02/03/2018  to account for ties, the statistical test only indicates if p-value <= alpha
+#             the exact test does not work for samples with different sizes
+
+# 03/03/2018  update of the Mann-Whitney exact test, now works for any sample size < 10
+#             above 10 data points the calculation time becomes very high
+
+# 04/03/2018  update of the Mann_whitney exact test, the script gets
+#             critical U statistics and cumulated W statistics frequencies
+#             from a text file to avoid calculating data on the fly
+
+# 05/03/2018  modified the way the legend is handled, removed "pad", set "loc" as "center left" and
+#             added the bbox_to_anchor parameter. To avoid cutting the legend, 
+#             added "bbox_inches ='tight'" argument when saving the figure.
+
+# 17/03/2018  addition of "Save as" options, the figure is stored in the new "Global" class
+#             and the "saveas" functions in the "Application" class calls it to save the figure
+
+# 20/03/2018  added statistical tests in a separate menu
+
+# 22/03/2018  optimization of figure display so figure is not cut
+
+# 30/04/2018 dunn.test method='bh' in R is 'fdr_bh' in Python
+#            process_csv can read index "Genotype" and "genotype"
+
+# 02/08/2018 add instructions to remove NaN from dataframe before processing
+#            the name of the index is not a problem now (cf 30/04/2018)
+
+# 06/08/2018 modified to comply with PEP8 (spaces around commas and '=')
+
+# 13/10/2018 Mann-Whitney test from scipy.stats is now used
+
+# 12/01/2019 remove error when closing window with 'Close' button
+#            user chooses where to save files (plots and test results)
+#            'Select all' button to select all genotypes for tests
+
+from tkinter import *
+from tkinter import filedialog
+import pandas as pd
+import numpy as np
+import matplotlib
+matplotlib.use("TkAgg")
+from matplotlib import pyplot as plt
+from scipy.stats import mannwhitneyu as mw
+import dunn_test 
+from itertools import combinations
+
+class Global(object):
+    """Class for global data"""
+    # dataframe data
+    struct = None
+    geno = None
+    col = None
+    number_rep = None
+    medians = None
+    medians_trans = None
+
+    # figure data
+    figure = None
+    ax = None
+
+    # file name from "filedialog.askopenfilename()"
+    file_name = None
+    
+    # counter for file extension
+    export_number_fig = 0
+    export_number_stats = 0
+    
+class MenuBar(Frame):
+    """drop-down menu bar"""
[…]
