+++ build/lib/pyms/__init__.py
@@ -1,0 +1,3 @@
+'''PyMS (Python for Mycorrhizal Symbiosis)'''
+
+__version__ = '0.8.1'
