### 2019-02-14 #1125 @jschnab
@arfon @aokad @johnlees
Hi, I added automatic replot when changing color and bar aggregation.

I made a pip package for PyPI but there is already a project called "pyms". I was going to rename my Github repository "pymsplot", can I keep the same JOSS submission? How do I update the link to the GitHub repository in the paper?
