### 2019-02-15 #1125 @arfon
> I made a pip package for PyPI but there is already a project called "pyms". I was going to rename my Github repository "pymsplot", can I keep the same JOSS submission? How do I update the link to the GitHub repository in the paper?

I can do this. Please go ahead and change the name and I'll update the details here.
