+++ build/lib/pyms/__main__.py
@@ -1,0 +1,2 @@
+from .pyms import main
+main()
