+++ pyms/__init__.py
@@ -1,3 +1,0 @@
-'''PyMS (Python for Mycorrhizal Symbiosis)'''
-
-__version__ = '0.8.1'
