+++ pyms_plot.egg-info/requires.txt
@@ -2,3 +2,4 @@ numpy
 scipy
 pandas
 matplotlib
+statsmodels
