+++ pyms_plot.egg-info/entry_points.txt
@@ -1,0 +1,3 @@
+[console_scripts]
+pyms = pyms.pyms:main
+
