+++ pyms_plot.egg-info/SOURCES.txt
@@ -1,0 +1,12 @@
+README.md
+setup.py
+pyms/__init__.py
+pyms/__main__.py
+pyms/dunn_test.py
+pyms/pyms.py
+pyms_plot.egg-info/PKG-INFO
+pyms_plot.egg-info/SOURCES.txt
+pyms_plot.egg-info/dependency_links.txt
+pyms_plot.egg-info/entry_points.txt
+pyms_plot.egg-info/requires.txt
+pyms_plot.egg-info/top_level.txt
