+++ pyms.egg-info/SOURCES.txt
@@ -1,10 +1,0 @@
-README.md
-setup.py
-pyms/__init__.py
-pyms/dunn_test.py
-pyms/pyms.py
-pyms.egg-info/PKG-INFO
-pyms.egg-info/SOURCES.txt
-pyms.egg-info/dependency_links.txt
-pyms.egg-info/requires.txt
-pyms.egg-info/top_level.txt
