### 2019-02-16 ff02be9
correct typo in README.md about installation
