### 2019-02-16 f71ad52
correct more typos in docs
