+++ docs/index.rst
@@ -14,20 +14,7 @@ Documentation
 Installation
 ------------
 
-PyMS is written in Python 3 and depends on [pandas 0.23](https://pandas.pydata.org/), [NumPy](http://www.numpy.org/), [matplotlib](https://matplotlib.org/), [SciPy](https://matplotlib.org/), [statsmodels](https://www.statsmodels.org) and [TkInter](https://docs.python.org/3/library/tk.html). You can install them by running on the command line::
-    python -m pip install <library name>
-
-After installing SciPy, in order for PyMS to be able to import ``scipy.stats``, run::
-    python -m pip install --upgrade scipy
-
-TkInter is traditionally bundled with Python. 
-
-Install PyMS through pip::
-    python -m pip install pyms-plot
-
-## Running PyMS
-
-Once PyMS is installed, you can run ``pyms`` on the command line. It may take a few seconds before the graphical interface appears.
+Please refer to the README file on the GibHub repository https://github.com/jschnab/pyms-plot.
 
 Input data and file format
 --------------------------
