### 2019-02-16 jschnab/pyms-plot#4 @jschnab
I added the PyMS package to PyPI under the name "pyms-plot"
