### 2019-02-16 #1125 @jschnab
@arfon I uploaded the PyMS package on PyPI under the name pyms-plot. I changed the name of the repository on GitHub to pyms-plot.
