### 2019-02-17 #1125 @jschnab
@arfon - I uploaded an archive on Zenodo the DOI is 10.5281/zenodo.2567459
By the way I updated the version of the software : 0.9.0
