### 2019-02-17 #1125 @arfon
@jschnab - it looks like there's an issue with one of the citations in the first paragraph?
