+++ paper/paper.bib
@@ -10,7 +10,7 @@
 @InProceedings{ SciPy,
   author    = { Eric Jones and Travis Oliphant and Pearu Peterson and others },
   title     = { {Scipy}: Open source scientific tools for {Python} },
-  year      = { 2001-- },
+  year      = { 2001 },
   url       = "http://www.scipy.org/",
   note      = {[Online; accessed 2019-01-13]}
 }
@@ -46,4 +46,11 @@
   Volume    = { 217 },
   Year      = { 2018 },
   doi       = { 10.1111/nph.14936 }
+}
+
+@InProceedings{ seabold,
+  title     = { Statsmodels: Econometric and statistical modeling with python },
+  author    = { Seabold, Skipper and Perktold, Josef },
+  booktitle = { 9th Python in Science Conference },
+  year      = { 2010 },
 }
