### 2019-02-17 7f67900
correct more mistakes in paper.md
