### 2019-02-17 #1125 @jschnab
@arfon - I corrected mistakes in paper.md
